adventureText Feature Adding
Added a bunch of stuff. Still need to go back in comment. Hidden items,
interactable objects (partial), initials for directions, more verbs can
be parsed, lots of extra rooms, items are changed, some vestiges of a
plot are to be found, some code optimizations here and there.

Kinda messed up as the thrid project Euler is due on Friday, and this is
due Wednesday, and I have not worked on Euler at all. That will take up
most of my time tomorrow.