Combining
Converted the midpoint from two separate functions that return a
variable to a single function that returns a Point object. I orginally
combined them into a single function that returns a list with each
coordinate value, but threw that out in favor of the current setup.

I also combined the Vertex and Point plotting functions into one
function that can be arged to plot either.